requirements.txt
python-telegram-bot[job-queue]==20.7
openai==1.14.0
python-dotenv==1.0.1
requests>=2.31.0
markdown>=3.5
